feat: add Windows compatibility and CI build (#1)
* chore: add max-external and max-patgen skills as submodules

Add .agents/skills/max-external and .agents/skills/max-patgen as git submodules for Max/MSP external building and patch generation skills.

* chore: update submodules to latest main after PR merges

max-external: fix version template, variable naming, quoting (#4)\nmax-patgen: fix typo, improve inlet indexing docs (#1)

* feat: add Art-Net viewer & controller UI in extras/

Add bbb.artnet.viewer.maxpat as a simple Art-Net DMX viewer and control UI patch.

Layout:
- Left: Controller section with target_ip selector, universe/subnet controls,
  and 8 channel number boxes (0-255) connected via pak to bbb.artnet.controller
- Right: Viewer section with bbb.artnet.node, universe/subnet controls,
  8-channel number display (unpack), and multislider (512ch) visual display

Update package-info.json with homepatcher field pointing to the viewer patch.
The patch appears in Max's Extras menu when the package is installed.

* feat: add Windows compatibility and CI build workflow

- Add net_compat.hpp: cross-platform abstraction for getifaddrs/GetAdaptersAddresses, WSAStartup, close/closesocket
- Refactor artnet_node_manager.hpp to use bbb::net::get_adapters() instead of raw getifaddrs
- Fix sACN controller/node: replace POSIX includes with net_compat, add const char* casts for Winsock API
- Add bbb::net::socket_valid() for cross-platform socket error checking
- Add bbb::net::recv_len_t (ssize_t/int) for recvfrom return type
- Add ws2_32/iphlpapi link libraries on WIN32 in root CMakeLists.txt
- Add .github/workflows/build.yml for macOS + Windows CI
- Add max-external-githubactions skill as submodule

* chore: update skill submodules to latest

max-external: MACOS_ONLY/WIN32_ONLY guards, pitfalls 9-11
max-external-githubactions: CI pitfalls, MACOS_ONLY setup guide

* fix: make asio clone conditional in CI workflow

asio directory may already exist from submodule checkout.
Check for README file before cloning to avoid failure.

* fix: Windows build errors from CI feedback

- Fix uniform_int_distribution<uint8_t> → int (MSVC rejects uint8_t)
- Fix OnLinkPrefixLength==0 shift UB in net_compat.hpp
- Add generate_version dependency to all project targets
  (MSBuild parallel builds could compile before version.h was generated)

* fix: Windows MSVC build errors + add MIT license and credits

Windows fixes:
- Set C++17 standard in root CMake (bit_by_bit requires C++11+)
- Replace static_cast<std::string>(symbol) with std::string(symbol) (MSVC)
- Use INVALID_SOCKET for SOCKET type initialization on Windows
- Fix uniform_int_distribution<uint8_t> → int (MSVC rejects char types)

Documentation:
- Add MIT LICENSE file
- Add acknowledgments: David Butler / imp.dmx
- Update build docs for cross-platform (macOS + Windows)

* fix: MSVC symbol-to-string conversion and __cplusplus detection

- Add /Zc:__cplusplus for MSVC (bit_by_bit checks __cplusplus macro)
- Use static_cast<const char*>(sym) for symbol→string conversion
  (MSVC cannot use operator std::string() for direct construction)
- Use direct assignment for symbol→m_bip_str (avoids ambiguity)

* fix: convert remaining static_cast<std::string>(symbol) in artnet.node

Use static_cast<const char*>(sym) for MSVC compatibility.

+++ source/bbb/sacn/sacn_packet.h
@@ -46,8 +46,8 @@ inline std::array<uint8_t, 16> generate_cid() {
     std::array<uint8_t, 16> cid;
     std::random_device rd;
     std::mt19937 gen(rd());
-    std::uniform_int_distribution<uint8_t> dist(0, 255);
-    for(auto& b : cid) b = dist(gen);
+    std::uniform_int_distribution<int> dist(0, 255);
+    for(auto& b : cid) b = static_cast<uint8_t>(dist(gen));
     cid[6] = (cid[6] & 0x0F) | 0x40;
     cid[8] = (cid[8] & 0x3F) | 0x80;
     return cid;
